Made a LoginHandling Component and cleaned up Login.jsx
