+++ src/pages/Login.jsx
@@ -1,22 +1,14 @@
-import React, { useState } from "react";
+import React from "react";
 import DocumentList from "../components/contexts/DocumentList";
+import LoginHandling from "../components/contexts/LoginHandling";
 import { useUser } from "../components/contexts/UserContext"; // Provides access to user context and related actions.
 
 const Login = () => {
   // Manages state for user credentials and form inputs.
-  const { user, handleLogin, handleLogout, error, documents } = useUser();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
+  const { user, handleLogin, handleLogout, error, documents } = useUser();  
 
-  // Handles user login on form submission.
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-    try {
-      await handleLogin(email, password);
-      console.log("Login successful!");
-    } catch (error) {
-      console.error("Login failed:", error.message);
-    }
+const handleLoginSubmit = async (email, password) => {
+    await handleLogin(email, password);
   };
 
   // Main render function for the login component.
@@ -35,25 +27,7 @@ const Login = () => {
           </div>
         </div>
       ) : (
-        <form onSubmit={handleSubmit}>
-          <label htmlFor="email">Email:</label>
-          <input
-            type="email"
-            id="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-          />
-          <label htmlFor="password">Password:</label>
-          <input
-            type="password"
-            id="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-          />
-          <button type="submit">Login</button>
-        </form>
+        <LoginHandling onSubmit={handleLoginSubmit} />
       )}
     </div>
   );
